Combine binaries, make subcommands
