+++ mobile_verifier/src/cli/mod.rs
@@ -1,0 +1,2 @@
+pub mod generate;
+pub mod server;
